Chore (Curve): Move multichain logic to EVM
