+++ rotkehlchen/chain/ethereum/modules/curve/accountant.py
@@ -2,11 +2,10 @@ import logging
 
 from rotkehlchen.chain.evm.accounting.interfaces import DepositableAccountantInterface
 from rotkehlchen.chain.evm.accounting.structures import EventsAccountantCallback
+from rotkehlchen.chain.evm.decoding.curve.constants import CPT_CURVE
 from rotkehlchen.history.events.structures.base import get_event_type_identifier
 from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
 from rotkehlchen.logging import RotkehlchenLogsAdapter
-
-from .constants import CPT_CURVE
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
